add new functions for item management module

+++ src/main/java/com/mry/service/CustomerService.java
@@ -5,8 +5,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.mry.config.SmsSetting;
+import com.mry.enums.CustomerStatus;
+import com.mry.enums.StoreStatus;
 import com.mry.model.Customer;
 import com.mry.repository.CustomerRepository;
+import com.mry.repository.StoreRepository;
 import com.mry.sms.SendSms;
 
 @Service
@@ -14,6 +17,8 @@ import com.mry.sms.SendSms;
 public class CustomerService {
 	@Resource
 	private CustomerRepository customerRepository;
+	@Resource
+	private StoreRepository storeRepository;
 	@Resource
 	private SmsSetting smsSetting;
 	@Resource
@@ -50,11 +55,53 @@ public class CustomerService {
 	
 	// 根据 手机号 删除一条用户信息
 	public int deleteCustomerByAccount(String account) {
-		return customerRepository.deleteCustomerByAccount(account);
+		Customer customer = customerRepository.getCustomerByAccount(account);
+		if(null != customer) {
+			// 删除用户信息
+			customerRepository.deleteCustomerByAccount(account);
+			// 修改门店的状态为注销状态
+			storeRepository.unsubscribeStore(customer.getId(), StoreStatus.CANCEL.getCode());
+			return 0;
+		}
+		return 1;
 	}
 	
-	// 根据 手机号 删除一条用户信息
+	// 根据 用户账号 删除一条用户信息
 	public int deleteCustomerByUserName(String userName) {
-		return customerRepository.deleteCustomerByUserName(userName);
+		Customer customer = customerRepository.getCustomerByUserName(userName);
+		if(null != customer) {
+			// 删除用户信息
+			customerRepository.deleteCustomerByUserName(userName);
+			// 修改门店的状态为注销状态
+			storeRepository.unsubscribeStore(customer.getId(), StoreStatus.CANCEL.getCode());
+			return 0;
+		}
+		return 1;
+	}
+	
+	// 根据 手机号 注销用户
+	public int unsubscribeByAccount(String account) {
+		Customer customer = customerRepository.getCustomerByAccount(account);
+		if(null != customer) {
+			// 修改用户的账号状态为注销状态
+			customerRepository.unsubscribeByAccount(account, CustomerStatus.CANCEL.getCode());
+			// 修改门店的状态为注销状态
+			storeRepository.unsubscribeStore(customer.getId(), StoreStatus.CANCEL.getCode());
+			return 0;
+		}
+		return 1;
+	}
+	
+	// 根据 用户账号 注销用户
+	public int unsubscribeByUserName(String userName) {
+		Customer customer = customerRepository.getCustomerByUserName(userName);
+		if(null != customer) {
+			// 修改用户的账号状态为注销状态
+			customerRepository.unsubscribeByUserName(userName, CustomerStatus.CANCEL.getCode());
+			// 修改门店的状态为注销状态
+			storeRepository.unsubscribeStore(customer.getId(), StoreStatus.CANCEL.getCode());
+			return 0;
+		}
+		return 1;
 	}
 }
